backend/requirements.txt
azure-core==1.35.1
azure-data-tables==12.7.0
azure-functions==1.23.0
azure-storage-blob==12.26.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
ecdsa==0.19.1
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.3
multidict==6.6.4
propcache==0.4.0
pyasn1==0.6.1
pycparser==2.23
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.21.0
